+++ Sources/StreamChat/Query/MessageSearchQuery.swift
@@ -56,13 +56,13 @@ public protocol AnyMessageSearchFilterScope {}
 
 public struct MessageSearchFilterScope: FilterScope, AnyMessageSearchFilterScope {}
 
-public extension FilterKey where Scope: AnyMessageSearchFilterScope {
+public extension FilterKey where Scope == MessageSearchFilterScope {
     static var text: FilterKey<Scope, String> { "text" }
     static var authorId: FilterKey<Scope, UserId> { "user.id" }
     static var hasAttachmentsOfType: FilterKey<Scope, AttachmentType> { "attachments.type" }
 }
 
-public extension Filter where Scope: AnyMessageSearchFilterScope {
+public extension Filter where Scope == MessageSearchFilterScope {
     static func queryText(_ text: String) -> Filter<Scope> {
         .query(.text, text: text)
     }