+++ Sources/StreamChat/Workers/UserListUpdater.swift
@@ -131,6 +131,7 @@ private extension UserPayload {
             lastActiveAt: lastActiveAt,
             teams: Set(teams),
             language: language.map(TranslationLanguage.init),
+            avgResponseTime: avgResponseTime,
             extraData: extraData
         )
     }